Update LayerActions.ts
Add a null-check and use typescript notation instead of javascript.

+++ csComp/services/layer/LayerActions.ts
@@ -138,9 +138,11 @@ module csComp.Services {
                 var reset = <IActionOption>{ title: 'Reset Layer', icon: 'reset' };
                 reset.callback = (layer: ProjectLayer, layerService: csComp.Services.LayerService) => {
                     console.log('Resetting layer: ' + layer.title);
-                    layer.data.features.forEach(function (f) {
-                        layer.layerSource.service.removeFeature(f)
-                    });
+                    if (layer.data && layer.data.features) {
+                        layer.data.features.forEach((f) => {
+                            layer.layerSource.service.removeFeature(f)
+                        });
+                    }
                     layer.data.features = [];
                     layer.layerSource.refreshLayer(layer);
                 };
